+++ docs/Changelog.md
@@ -8,7 +8,9 @@
 
 ### `3.8.0`
 
- * Configurable opcode/class mapping (@sirn-se)
+ * Configurable opcode/class mapping using OpcodeRegistry (@sirn-se)
+ * onMessage() listener, triggered on all messages (@sirn-se)
+ * Connections container class (@sirn-se)
  * Dispach onDisconnect on remote connection close (@srebb, @sirn-se)
  * Exceptions: ControlInterface on CloseException, ReconnectException (@sirn-se)
  * Exceptions: HandlerLevelInterface on BadUriException, ClientException, RunnerException, ServerException (@sirn-se)